requirements.txt
cat > requirements.txt << 'EOF'
Flask==2.3.3
flask-cors==4.0.1
requests==2.31.0
gunicorn==22.0.0
python-dotenv==1.0.1
authlib==1.3.0
flask-limiter==3.5.0
redis==5.0.1
EOF